Add more abstraction
The code contains a few very long methods.

Let's abstract out the logic to make the code neater.
